Moved glue.qt.mime to glue.core.qt.mime
