+++ glue/core/qt/mime.py
