+++ glue/core/qt/tests/test_data_collection_model.py
@@ -4,7 +4,7 @@ from glue.external.qt.QtCore import Qt
 from glue.core import DataCollection, Data
 
 from ..data_collection_model import DataCollectionModel
-from glue.qt.mime import LAYERS_MIME_TYPE
+from glue.core.qt.mime import LAYERS_MIME_TYPE
 
 
 class TestDataCollectionModel(object):
