+++ glue/viewers/common/qt/data_viewer.py
@@ -7,7 +7,7 @@ from glue.external.qt import QtGui
 from glue.core.application_base import ViewerBase
 from glue.core.qt.layer_artist_model import QtLayerArtistContainer, LayerArtistView
 from glue.qt import get_qapp
-from glue.qt.mime import LAYERS_MIME_TYPE, LAYER_MIME_TYPE
+from glue.core.qt.mime import LAYERS_MIME_TYPE, LAYER_MIME_TYPE
 from glue.utils.qt import set_cursor
